Yay! This is my first swift program.
I am happy that I got two classes working together as one is a delegate of the other. If the function throws, then ViewController catches the error and passes it to BAM which prints the force unwrapped optional, cashCab. 
In my first repository, codeja, I want my program to also have a delegate class. The ViewController will play a song and start a slider and label to keep track of the time; at the end of the song the delegate will stop the slider and label from progressing. Also, if the ViewController throws an error, the delegate will tell me why.